feat: Add NextButton component
This commit adds a new component called NextButton. The NextButton component is responsible for rendering a button with customizable text and style. It includes an onPress event handler and supports disabling the button. The styles for the component are defined in the StyleSheet.create function.

The addition of the NextButton component will improve the consistency and reusability of next buttons throughout the application.